Save/restore mowing progress across service/container restarts
Records the current plan progress in a bag file and restores on restart if
exactly the same plan is being mowed.

A digest of all of the poses in the plan is used to determine if the plan is
identical. This should work regardless of why the plan changed e.g if any
params that affect slicing change, or if the slicer logic changes.